==> Spatial_comparison_and_Isolation_index/01_scripts/04_cetesb_rds_to_csv.R <==
# Location of input files and output directory
cetesb_data_file  <- "02_data/raw_data/cetesb_010320_140420.Rds"


# Loading pollutant information for all cetesb stations
cetesb_data <- readRDS(cetesb_data_file)


# Function to write station data in csv

WriteStationData<- function(df){
  station_name <- unique(na.omit(df$est))
  station_name <- iconv(station_name, from = "UTF-8", to = "ASCII//TRANSLIT") # Change spetial characters
  station_name <- gsub(" ", "_", station_name) # Removing empty spaces in station_name
  
  path <- "./02_data/cetesb_data/"
  file_name <- paste0(path, station_name, "_010320-140420.csv")
  write.table(df, file_name, quote = F, row.names = F, sep = ',')
}


lapply(cetesb_data, WriteStationData)
